Don't update namedtuple
This is a major blocker for deployment. If we update this, then we have
to update the API to add the extra field. If we update the API, we break
every existing groupy client.

In short: everything is terrible. This will at least let us add the
audited flag to groups, in a backwards compatible way.