Revert "Change User.role_user to User.service_account"
This reverts commit 123fc47b98a52885935f73598bc3dee3fab1372e.
This is being done due to an upstream change in Grouper, where service
accounts are no longer exposed in the API as if they are users, but
rather as separate first class objects. For more information see the
commit (e9cdb271e9f2bc48757f6b701f5d27b0ed28e87a) in Grouper, as well as
Grouper pull request #266.