Support new service accounts, add password support
This is a slightly hacky approach that just exposes the new
service account metadata if present and rewrites service account
permissions to `None` the fields that they don't have.  A better
long-term approach would polymorphism and return ServiceAccount
objects alongside User objects, but that raises various
complications and is deferred for future work.

While updating the list of supported fields, add support for
passwords to match the current graph API.