Make flake8 and pytest dependencies conditional
Rather than unconditionally adding flake8, flake8-import-order,
and pytest-runner to setup_requires, only add them if invoked
with arguments that would require them.  This allows those
dependencies to be omitted in environments that only care about
installing and running the code.

Ensure the dependencies for flake8 and test commands are always
included in tests_require so that they'll be picked up by CI.

Fixes #22