Fix flake8 warnings about imports, bump versions
The latest version of flake8-import-order is pickier and doesn't
like the Dropbox house style.  Switch to smarkets, which is closer
to house style but without the blank line between `import` and
`from` directives, and remove the blank line.

Bump the versions of flake8 and flake8-import-order so that we'll
get consistent test results.