Bump requirements, switch to >=
groupy is pretty simple and hopefully won't break with new versions,
so be optimistic and use >= for all requirements so that they don't
have to be kept in sync with merou.

Bump the tornado, black, mypy, flake8, and flake8-import-order
versions to match merou.

Rename requirements-dev3.txt to requirements-dev.txt and
requirements-dev.txt to requirements-dev2.txt to reflect the
default development environment is now Python 3.

Fix some type errors with the latest version of mypy.